move getmodel
